+++ src/literalizer/languages/dhall.py
@@ -48,6 +48,7 @@ from literalizer._language import (
     DictFormatConfig,
     FloatSpecialsMixin,
     HeterogeneousBehavior,
+    IdentifierCase,
     LanguageCls,
     OrderedMapFormatConfig,
     SequenceFormatConfig,
@@ -549,6 +550,11 @@ class Dhall(metaclass=LanguageCls):
         """
 
     heterogeneous_strategies = HeterogeneousStrategies
+
+    identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
+        IdentifierCase.SNAKE,
+        IdentifierCase.CAMEL,
+    )
 
     validate_spec_for_data = no_validate_spec_for_data