+++ tests/integration/cases/call_ref_args_converted_nonsnake/Clojure_call.clj
@@ -1,0 +1,13 @@
+(defn process [& _args] nil)
+(def my-var [
+    1
+    2
+    3
+])
+(def my-other [
+    4
+    5
+    6
+])
+(process :data my-var :count 42)
+(process :data my-other :count 7)