+++ tests/integration/cases/call_ref_args_converted_whole/FSharp_call.fs
@@ -1,0 +1,12 @@
+module Check
+
+type Val =
+    | FInt of int64
+    | FList of Val list
+let process (_data: obj) : obj = null
+let myVar: Val = FList [
+    FInt 1L;
+    FInt 2L;
+    FInt 3L
+]
+process(myVar)