+++ tests/integration/cases/call_ref_args_converted_whole/Java_call.java
@@ -1,0 +1,11 @@
+class Check {
+static Object process(Object... args) { return null; }
+    public static void check() {
+var myVar = new int[]{
+    1,
+    2,
+    3
+};
+process(myVar);
+    }
+}